book_project/backend_rag_service/requirements.txt
fastapi
uvicorn
google-generativeai
qdrant-client
python-dotenv
asyncpg  # for Neon Serverless Postgres
sqlalchemy
httpx
pytest-asyncio
